Update app.js line 59
Choosing to use arrow keys, such as ArrowRight, ArrowUp, ArrowLeft, and ArrowDown, instead of numbers for keyboard actions, enhances a more intuitive and natural interface. Arrow keys are universally recognized as directional symbols, making it easier for users to understand movement within the interface. This choice improves the user experience by providing a more consistent and easily memorable navigation. Additionally, using arrow keys contributes to user familiarity, as it is a common convention in software and games, promoting more efficient and comfortable interaction.😁